refactor(daemon): move claude-runtime into a neutral runtime-defs leaf
`src/acp/claude-runtime.ts` only depends on `config/config-schema` and
`node:path`, yet living under `src/acp/` forced three `runtimes/` modules
(config-caps, model-provider-config, launch-policy) to import backwards into
`acp/`. Move it to `src/runtime-defs/claude-runtime.ts` and update importers.

Pure move — no behavior change.
